% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaseddata_datagenerator.R
\name{biaseddata_datagenerator}
\alias{biaseddata_datagenerator}
\title{Biased Data - generator}
\usage{
biaseddata_datagenerator(n1, n2, theta1_star, theta2_star)
}
\description{
Generate synthetic dataset from the biased data model
}
