% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaseddata_module1.R
\name{biaseddata_module1}
\alias{biaseddata_module1}
\title{Biased Data - module 1}
\usage{
biaseddata_module1()
}
\description{
Generate a module object, in the form of a list, that contains all
we need from module 1: a function to evaluate the prior, a function to evaluate the log-likelihood,
...
}
