% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaseddata_module2.R
\name{biaseddata_module2}
\alias{biaseddata_module2}
\title{Biased Data - module 2}
\usage{
biaseddata_module2()
}
\description{
Generate a module object, in the form of a list, that contains all
we need from module 2: a function to evaluate the prior, a function to evaluate the log-likelihood,
...
}
