% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainlist_to_dataframe.R
\name{chainlist_to_dataframe}
\alias{chainlist_to_dataframe}
\title{Convert list of chains to data.frame}
\usage{
chainlist_to_dataframe(chains_list)
}
\description{
to handle the output of metropolishastings
}
