% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolishastings.R
\name{metropolishastings}
\alias{metropolishastings}
\title{Metropolis Hastings algorithms}
\usage{
metropolishastings(target, tuning_parameters)
}
\description{
with multiple chains and adaptive proposal
}
