% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plummer_module2.R
\name{plummer_module2}
\alias{plummer_module2}
\title{Plummer Model - module 2}
\usage{
plummer_module2()
}
\description{
Generate a module object, in the form of a list, that contains all
we need from module 1: a function to evaluate the prior, a function to evaluate the log-likelihood,
...
}
