% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plummer_module2_conditional.R
\name{plummer_module2_conditional}
\alias{plummer_module2_conditional}
\title{Plummer Model - module 2 posterior conditional on theta1 in c++}
\usage{
plummer_module2_conditional(theta1, theta2s, ncases, Npop_normalized)
}
